DWORD should now be 4 bytes on all OSs

+++ XboxInternals/winnames.h
@@ -1,11 +1,11 @@
 #ifndef WINNAMES_H
 #define WINNAMES_H
 
-typedef unsigned int INT24;
+typedef signed int INT24;
 typedef unsigned char BYTE;
 typedef short INT16;
 typedef unsigned short WORD;
-typedef unsigned long DWORD;
+typedef unsigned int DWORD;
 typedef int INT32;
 typedef long long INT64;
 typedef unsigned long long UINT64;
